+++ js/string-inspection.js
@@ -1,0 +1,4 @@
+// Проверка строки
+const inspectString = (string, maxWidth = 140) => string.length < maxWidth;
+
+export {inspectString};
